SHA512 hash as book identifier instead of fake ISBN
ISBNs are meant for books which actually have been registered in one of the
organisms which assign ISBNs. It is more sensible to use a hash based on the
content of the downloaded HTML source files and the contents of "xtool"
itself. The later means that modifying "xtool" (and hence the way the
converted data looks) will generate a new book identifier.